calendar is now clickable and returns the selected date
