--- my-coaching-pal/pages/coach/[username]/[playerUsername]/index.js
@@ -1,10 +1,0 @@
-import { useRouter } from "next/router";
-import styles from "../../../../styles/Home.module.css"
-
-export default function CoachPlayerPage(){
-    const router = useRouter();
-    const playerUsername = router.query.playerUsername;
-    return(
-        <h2>{playerUsername}</h2>
-    );
-}
